Add files via upload
In Monte Carlo simulations, random sampling and probabilistic rules are employed to approximate and analyze complex systems. In the case of the Ising Model, the Monte Carlo simulation is used to update the spins at each step based on the current state and probabilistic rules. This allows for the exploration of different spin configurations and the study of the system’s properties at different temperatures.
Interaction: The spins interact with their nearest neighbors. The strength of this interaction is determined by the interaction parameter interaction passed to the IsingModel3d constructor; the interaction parameter is denoted as self.j
Simulation Step: The simulation evolves the system over time by updating the spins according to a probabilistic rule. In the code, this is implemented in the make_simulation_step method. Each spin is considered in a sequential manner, and the probability of flipping its state is determined based on the energy change and the temperature. The spin is flipped with a certain probability calculated using the Metropolis algorithm.